+++ packages/graphql-tower-storage-s3/src/__tests__/index.test.js
@@ -7,7 +7,19 @@ jest.mock('graphql-tower-unique', () => () => 'UNIQUE_GUID');
 
 const formData = jest.fn();
 global.FormData = () => ({ append: formData });
-global.fetch = jest.fn();
+
+const xmlHttpRequest = jest.fn();
+global.XMLHttpRequest = () => {
+  let req;
+  const obj = {
+    open: (...args) => { req = args; },
+    onload: () => {},
+    onerror: () => {},
+    send: () => xmlHttpRequest(...req, obj).then(obj.onload, obj.onerror),
+    upload: {},
+  };
+  return obj;
+};
 
 describe('storage s3', () => {
   const storage = new StorageS3({ STORAGE_URL: 's3://username:secret_key@ap-northeast-1/graphql-tower' });
@@ -136,14 +148,36 @@ describe('storage s3', () => {
 
   describe('upload', () => {
     it('successfully upload', async () => {
-      upload(storage.generateTemporaryCredentials(), 'FILE_PATH');
+      xmlHttpRequest.mockReturnValueOnce(Promise.resolve());
+      await upload(storage.generateTemporaryCredentials(), 'FILE_PATH');
       expect(formData).toHaveBeenCalledWith('file', 'FILE_PATH');
       expect(formData).toHaveBeenCalledWith('policy', expect.anything());
+      expect(formData).toHaveBeenCalledWith('success_action_status', '201');
       expect(formData).toHaveBeenCalledWith('x-amz-algorithm', 'AWS4-HMAC-SHA256');
-      expect(formData).toHaveBeenCalledWith('x-amz-credential', expect.anything());
+      expect(formData).toHaveBeenCalledWith('x-amz-credential', expect.stringContaining('username/'));
       expect(formData).toHaveBeenCalledWith('x-amz-date', expect.anything());
       expect(formData).toHaveBeenCalledWith('x-amz-signature', expect.anything());
-      expect(fetch).toHaveBeenCalledTimes(1);
+      expect(xmlHttpRequest).toHaveBeenCalledWith('POST', 'https://graphql-tower.s3.amazonaws.com', expect.anything());
+      expect(xmlHttpRequest).toHaveBeenCalledTimes(1);
+
+      const request = xmlHttpRequest.mock.calls[0][2];
+      request.upload.onprogress({ lengthComputable: true, loaded: 50, total: 100 });
+    });
+
+    it('when network error', async () => {
+      xmlHttpRequest.mockReturnValueOnce(Promise.reject());
+      await expect(upload(storage.generateTemporaryCredentials(), 'FILE_PATH'))
+        .rejects.toEqual(new Error('failed to upload: network error'));
+    });
+
+    it('onprogress', async () => {
+      const onprogress = jest.fn();
+      xmlHttpRequest.mockReturnValueOnce(Promise.resolve());
+      await upload(storage.generateTemporaryCredentials(), 'FILE_PATH', onprogress);
+
+      const request = xmlHttpRequest.mock.calls[0][2];
+      request.upload.onprogress({ lengthComputable: true, loaded: 50, total: 100 });
+      expect(onprogress).toHaveBeenCalledWith(0.5);
     });
   });
 });